Fix the build-coverage and further optimize scss

+++ build.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 # clone source
-git clone --depth 1 --branch v5.0.1 https://github.com/twbs/bootstrap.git ./var/build/bootstrap
-rm -rf ./var/build/bootstrap/.git
+if [ ! -d "./var/build/bootstrap" ]; then
+  git clone --depth 1 --branch v5.0.1 https://github.com/twbs/bootstrap.git ./var/build/bootstrap
+  rm -rf ./var/build/bootstrap/.git
+fi
 
 # enter directory
 cd ./var/build/bootstrap || exit 1
@@ -10,6 +12,8 @@ npm i
 
 # replace source style
 cp -f ../../../res/style/bootstrap.scss ./scss/bootstrap.scss
+cp -f ../../../res/style/_forms.scss ./scss/_forms.scss
+cp -f ../../../res/style/_helpers.scss ./scss/_helpers.scss
 
 # replace source script
 cp -f ../../../res/script/index.umd.js ./js/index.umd.js
@@ -26,5 +30,7 @@ cp -f ./dist/js/bootstrap.min.js.map ../../../public/res/script/bootstrap.bundle
 # exit directory
 cd ../../../
 
+sed -i -e 's/bootstrap\.min\.js\.map/bootstrap\.bundle\.min\.js\.map/g' ./public/res/script/bootstrap.bundle.min.js
+
 # delete source
-rm -rf ./var/build/bootstrap
+#rm -rf ./var/build/bootstrap
